feat: The end is near. (unofficial)

+++ SimpleBinaryVCS/Utils/FileHandlerTool.cs
@@ -1,25 +1,77 @@
 ﻿using SimpleBinaryVCS.DataComponent;
 using SimpleBinaryVCS.Interfaces;
+using SimpleBinaryVCS.Model;
 using System.IO;
 
 namespace SimpleBinaryVCS.Utils
 {
     public class FileHandlerTool
     {
-        public void HandleFile(IProjectData fileData, DataChangedState state)
+        public void ApplyFileChanges(List<ChangedFile> ChangedFile)
+        {
+            foreach (ChangedFile file in ChangedFile)
+            {
+                HandleData(file.SrcFile, file.DstFile, file.DataState);
+            }
+        }
+
+        public void HandleData(IProjectData dstData, DataChangedState state)
+        {
+            if (dstData.DataType == ProjectDataType.File)
+            {
+                HandleFile(null, dstData.DataAbsPath, state);
+            }
+            else
+            {
+                HandleDirectory(null, dstData.DataAbsPath, state);
+            }
+        }
+        public void HandleData(IProjectData? srcData, IProjectData dstData, DataChangedState state)
+        {
+            if (dstData.DataType == ProjectDataType.File)
+            {
+                HandleFile(srcData?.DataAbsPath, dstData.DataAbsPath, state);
+            }
+            else
+            {
+                HandleDirectory(srcData?.DataAbsPath, dstData.DataAbsPath, state);
+            }
+        }
+        public void HandleData(string? srcPath, string dstPath, ProjectDataType type, DataChangedState state)
+        {
+            if (type == ProjectDataType.File)
+            {
+                HandleFile(srcPath, dstPath, state);
+            }
+            else
+            {
+                HandleDirectory(srcPath, dstPath, state);
+            }
+        }
+        public void HandleData(string? srcPath, IProjectData dstData, DataChangedState state)
+        {
+            if (dstData.DataType == ProjectDataType.File)
+            {
+                HandleFile(srcPath, dstData.DataAbsPath, state);
+            }
+            else
+            {
+                HandleDirectory(srcPath, dstData.DataAbsPath, state);
+            }
+        }
+        public void HandleDirectory(string? srcPath, string dstPath, DataChangedState state)
         {
             try
             {
-                switch (state)
+                if ((state & DataChangedState.Deleted) != 0)
                 {
-                    case DataChangedState.Added:
-                        break;
-                    case DataChangedState.Modified: 
-                        break;
-                    case DataChangedState.Deleted: 
-                        break;
-                    default:
-                        break;
+                    if (Directory.Exists(dstPath))
+                        Directory.Delete(dstPath, true);
+                }
+                else
+                {
+                    if (!Directory.Exists(dstPath))
+                        Directory.CreateDirectory(dstPath);
                 }
             }
             catch (Exception ex)
@@ -27,63 +79,19 @@ namespace SimpleBinaryVCS.Utils
                 MessageBox.Show(ex.Message);
             }
         }
-        public void HandleFile(IProjectData srcData, IProjectData dstData, DataChangedState state)
+        public void HandleFile(string? srcPath, string dstPath, DataChangedState state)
         {
             try
             {
-                switch (state)
+                if ((state & DataChangedState.Deleted) != 0)
                 {
-                    case DataChangedState.Added:
-                        break;
-                    case DataChangedState.Modified:
-                        break;
-                    case DataChangedState.Deleted:
-                        break;
-                    default:
-                        break;
-                }
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
-            }
-        }
-        public void HandleFile(string srcPath, string dstPath, DataChangedState state)
-        {
-            try
-            {
-                switch (state)
-                {
-                    case DataChangedState.Added:
-                        break;
-                    case DataChangedState.Modified:
-                        break;
-                    case DataChangedState.Deleted:
-                        break;
-                    default:
-                        break;
-                }
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
-            }
-        }
-        public void HandleFile(string dstPath, DataChangedState state)
-        {
-            try
-            {
-                switch (state)
-                {
-                    case DataChangedState.Added:
-                        break;
-                    case DataChangedState.Modified:
-                        break;
-                    case DataChangedState.Deleted:
+                    if (File.Exists(dstPath))
                         File.Delete(dstPath);
-                        break;
-                    default:
-                        break;
+                }
+                else
+                {
+                    if (srcPath == null) throw new ArgumentNullException(nameof(srcPath));
+                    File.Copy(srcPath, dstPath, true);
                 }
             }
             catch (Exception ex)
